Move Exported and Temp deck to shared.
