+++ Server/src/server/dataCenter/models/card/ExportedDeck.java
@@ -1,35 +1,9 @@
 package server.dataCenter.models.card;
 
-import shared.models.card.Card;
+import shared.models.card.BaseExportedDeck;
 
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-
-public class ExportedDeck {
-    private final String name;
-    private String heroName;
-    private final HashMap<String, Integer> otherCards = new HashMap<>();
-
+public class ExportedDeck extends BaseExportedDeck<ServerCard, Deck> {
     public ExportedDeck(Deck deck) {
-        name = deck.getName();
-        if (deck.getHero() != null) {
-            heroName = deck.getHero().getName();
-        }
-        for (Card card : deck.getCards()) {
-            otherCards.merge(card.getName(), 1, Integer::sum);
-        }
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getHeroName() {
-        return heroName;
-    }
-
-    public Map<String, Integer> getOtherCards() {
-        return Collections.unmodifiableMap(otherCards);
+        super(deck);
     }
 }
