+++ Server/src/server/dataCenter/models/card/TempDeck.java
@@ -1,34 +1,9 @@
 package server.dataCenter.models.card;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import shared.models.card.BaseTempDeck;
 
-public class TempDeck {
-    private final String deckName;
-    private String heroId;
-    private final List<String> cardIds = new ArrayList<>();
-
+public class TempDeck extends BaseTempDeck<ServerCard, Deck> {
     public TempDeck(Deck deck) {
-        this.deckName = deck.getName();
-        if (deck.getHero() != null) {
-            this.heroId = deck.getHero().getCardId();
-        }
-
-        for (ServerCard card : deck.getCards()) {
-            this.cardIds.add(card.getCardId());
-        }
-    }
-
-    public String getDeckName() {
-        return deckName;
-    }
-
-    String getHeroId() {
-        return heroId;
-    }
-
-    List<String> getCardIds() {
-        return Collections.unmodifiableList(cardIds);
+        super(deck);
     }
 }
